+++ src/main/java/dp/futsal/database/Results6Repo.java
@@ -4,7 +4,7 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface Results6Repo extends JpaRepository<Results6, Long> {
+public interface Results6Repo extends JpaRepository<Results6, Integer> {
 
     public List<Results6> findByMDay(int mDay);
