+++ src/main/java/dp/futsal/database/Results9Repo.java
@@ -4,7 +4,7 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface Results9Repo extends JpaRepository<Results9, Long> {
+public interface Results9Repo extends JpaRepository<Results9, Integer> {
 
     public List<Results9> findByMDay(int mDay);
